training/requirements.txt
torch>=2.0.0
torchvision>=0.15.0
numpy==1.26
Pillow>=9.0.0
matplotlib>=3.5.0
arabic-reshaper>=3.0.0
python-bidi>=0.4.2